requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
httpx>=0.26.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.0.0
